improving vector2d struct

+++ src/engine/Vector2d.hpp
@@ -3,7 +3,7 @@
 struct Vector2D {
     float x;
     float y;
-    explicit Vector2D(int x = 0, int y = 0) : x(x), y(y) {}
+    explicit Vector2D(float x = 0, float y = 0) : x(x), y(y) {}
 
     Vector2D& operator=(const Vector2D& a) {
         x = a.x;
@@ -31,11 +31,11 @@ struct Vector2D {
         return *this;
     }
 
-    Vector2D operator*(const int a) const {
+    Vector2D operator*(const float a) const {
         return Vector2D(x * a, y * a);
     }
 
-    Vector2D operator*=(const int a) {
+    Vector2D operator*=(const float a) {
         x = x * a;
         y = y * a;
         return *this;
